add tests for content attribute
